+++ src/assets/Clock.tsx
@@ -1,0 +1,25 @@
+interface PropsType {
+  size?: number;
+  onClick?: () => void;
+  className?: string;
+}
+
+export const Clock = ({
+  size = 24,
+  onClick,
+  className
+}: PropsType) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      viewBox="0 0 24 24"
+      fill="none"
+      onClick={onClick}
+      className={`cursor-pointer ${className}`}
+    >
+      <path d="M12 7V12H17M12 21C7.02944 21 3 16.9706 3 12C3 7.02944 7.02944 3 12 3C16.9706 3 21 7.02944 21 12C21 16.9706 16.9706 21 12 21Z" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
+    </svg>
+  )
+}
