+++ src/assets/Timer.tsx
@@ -1,0 +1,25 @@
+interface PropsType {
+  size?: number;
+  onClick?: () => void;
+  className?: string;
+}
+
+export const Timer = ({
+  size = 24,
+  onClick,
+  className
+}: PropsType) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      viewBox="0 0 24 24"
+      fill="none"
+      onClick={onClick}
+      className={`cursor-pointer ${className}`}
+    >
+      <path d="M12 13V9M21 6L19 4M10 2H14M12 21C7.58172 21 4 17.4183 4 13C4 8.58172 7.58172 5 12 5C16.4183 5 20 8.58172 20 13C20 17.4183 16.4183 21 12 21Z" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
+    </svg>
+  )
+}
